Remove redundant C-style casts
- Remove redundant cast on line 462 where n is already size_t
- Fix similar pattern on line 393 using 1UL instead of C-style cast

+++ posts/gups/gups.cu
@@ -390,7 +390,7 @@ int main(int argc, char* argv[])
     }
   }
 
-  size_t n_shmem = (size_t)1 << logn_shmem;
+  size_t n_shmem = 1UL << logn_shmem;
 
   int ndev;
   CUDA_RT_CALL(cudaGetDeviceCount(&ndev));
@@ -459,7 +459,7 @@ int main(int argc, char* argv[])
   }
 
   size_t n = 1UL << logn;
-  size_t working_set = (size_t)n * 100 / occupancy;
+  size_t working_set = n * 100 / occupancy;
   if (accesses_per_elem == -1) {
     if (test_type == UPDATE || test_type == UPDATE_NO_LOOP
         || test_type == READ_WRITE) {
